+++ source/Messaging.Infrastructure/IncomingMessages/MessageQueueDispatcher.cs
@@ -19,6 +19,7 @@ using System.Transactions;
 using Azure.Messaging.ServiceBus;
 using Messaging.Application.Configuration;
 using Messaging.Application.IncomingMessages;
+using Messaging.Application.IncomingMessages.RequestChangeOfSupplier;
 using Messaging.CimMessageAdapter.Messages;
 using Messaging.Infrastructure.Configuration.Serialization;
